feat(api): block api endpoints if no api key

+++ app/pages/api/audio.ts
@@ -5,41 +5,59 @@ import OpenAI from "openai";
 import { tmpdir } from "os";
 import { join } from "path";
 import { promises as fsPromises } from "fs";
+import { verifyKey } from "@unkey/api";
 
 export const config = {
-	api: {
-		bodyParser: {
-			sizeLimit: "20mb",
-		},
-	},
+  api: {
+    bodyParser: {
+      sizeLimit: "20mb",
+    },
+  },
 };
 
+
 export default async function handler(
-	req: NextApiRequest,
-	res: NextApiResponse
+  req: NextApiRequest,
+  res: NextApiResponse
 ) {
-	if (req.method === "POST") {
-		const { file } = req.body;
-		const base64Data = file.split(";base64,").pop();
-		const tempFilePath = join(tmpdir(), `upload_${Date.now()}.mp3`);
-		await fsPromises.writeFile(tempFilePath, base64Data, {
-			encoding: "base64",
-		});
+  const header = req.headers.authorization;
+  if (!header) {
+    return new Response("No Authorization header", { status: 401 });
+  }
+  const token = header.replace("Bearer ", "");
+  const { result, error } = await verifyKey(token);
 
-		const apiKey = process.env.OPENAI_API_KEY;
-		const openai = new OpenAI({ apiKey, dangerouslyAllowBrowser: true });
-		const transcription = await openai.audio.transcriptions.create({
-			file: fs.createReadStream(tempFilePath),
-			model: "whisper-1",
-		});
+  if (error) {
+    console.error(error.message);
+    return new Response("Internal Server Error", { status: 500 });
+  }
 
-		// Clean up the temporary file
-		await fsPromises.unlink(tempFilePath);
+  if (!result.valid) {
+    // do not grant access
+    return new Response("Unauthorized", { status: 401 });
+  }
+  if (req.method === "POST") {
+    const { file } = req.body;
+    const base64Data = file.split(";base64,").pop();
+    const tempFilePath = join(tmpdir(), `upload_${Date.now()}.mp3`);
+    await fsPromises.writeFile(tempFilePath, base64Data, {
+      encoding: "base64",
+    });
 
-		res.status(200).json({ text: transcription.text });
-	} else {
-		// Handle any non-POST requests
-		res.setHeader("Allow", ["POST"]);
-		res.status(405).end(`Method ${req.method} Not Allowed`);
-	}
+    const apiKey = process.env.OPENAI_API_KEY;
+    const openai = new OpenAI({ apiKey, dangerouslyAllowBrowser: true });
+    const transcription = await openai.audio.transcriptions.create({
+      file: fs.createReadStream(tempFilePath),
+      model: "whisper-1",
+    });
+
+    // Clean up the temporary file
+    await fsPromises.unlink(tempFilePath);
+
+    res.status(200).json({ text: transcription.text });
+  } else {
+    // Handle any non-POST requests
+    res.setHeader("Allow", ["POST"]);
+    res.status(405).end(`Method ${req.method} Not Allowed`);
+  }
 }
